+++ palette.py
@@ -5,6 +5,7 @@ LIGHT for light theme
 DARK for dark theme
 """
 
-LIGHT_BACKGROUND = [255, 255, 255, 1]
+LIGHT_BACKGROUND = [1, 1, 1, 1]
 LIGHT_TEXT = [0, 0, 0, 1]
-LIGHT_SELECTION = [150, 220, 240, 1]
+LIGHT_BLUE = [0.6, 0.85, 0.9, 1]
+LIGHT_LINES = [0.5, 0.5, 0.5, 1]